code refactoring to make code reusable for both Open Channel and Group Channel
